// frontend/src/api/model/authorizationState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type AuthorizationState = typeof AuthorizationState[keyof typeof AuthorizationState];


export const AuthorizationState = {
  REQUESTED: 'REQUESTED',
  PENDING: 'PENDING',
  AUTHORIZED: 'AUTHORIZED',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED',
} as const;