// frontend/src/api/model/authorizationType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type AuthorizationType = typeof AuthorizationType[keyof typeof AuthorizationType];


export const AuthorizationType = {
  ADMISSION: 'ADMISSION',
  EXTENSION: 'EXTENSION',
  PROCEDURE: 'PROCEDURE',
  TRANSFER: 'TRANSFER',
  OTHER: 'OTHER',
} as const;