// frontend/src/api/model/careTeamRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type CareTeamRole = typeof CareTeamRole[keyof typeof CareTeamRole];


export const CareTeamRole = {
  ATTENDING_PHYSICIAN: 'ATTENDING_PHYSICIAN',
  SPECIALIST: 'SPECIALIST',
  RESIDENT: 'RESIDENT',
  NURSE: 'NURSE',
  OTHER: 'OTHER',
} as const;